playlist and song recommendation

+++ home/models.py
@@ -6,13 +6,15 @@ class Songs(models.Model):
     link_of_song  = models.URLField(max_length = 100)
     artist = models.CharField(max_length=100,null = True,blank = True)
     genre =  models.CharField(max_length = 100)
-    angry = models.PositiveIntegerField()
-    disgust = models.PositiveIntegerField()
-    fear = models.PositiveIntegerField()
-    happy = models.PositiveIntegerField()
-    neutral = models.PositiveIntegerField()
-    sad = models.PositiveIntegerField()
-    surprise = models.PositiveIntegerField()
+    angry = models.DecimalField(max_digits=5, decimal_places=2)
+    disgust = models.DecimalField(max_digits=5, decimal_places=2)
+    fear = models.DecimalField(max_digits=5, decimal_places=2)
+    happy = models.DecimalField(max_digits=5, decimal_places=2)
+    neutral =models.DecimalField(max_digits=5, decimal_places=2)
+    sad = models.DecimalField(max_digits=5, decimal_places=2)
+    surprise = models.DecimalField(max_digits=5, decimal_places=2)
     
     def __str__(self):
         return  f'{self.name}  by  {self.artist}'
+
+
